Makefile: Fix wrong prefix when compiling assembly source files
When compiling with style like "CC foo.o" assembly source files are
being used with wrong prefix( which is CC). This may result confusion of which
object file was assembly and which was C.

Project that use this design for compiling like linux kernel use "AS" to
indicate the object file being compiled is assembly.

Chnage the prefix for assembly files to "AS" to indicate object being
assembled is assembly file.

Signed-off-by: khalid Ali <khaliidcaliy@gmail.com>